feat: introduce ProfileManager
The ProfileManager replaces the few methods in AccountManager that were
concerned with fetching remote profiles and updating our own profile.
